backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cicd-orchestrator"
version = "0.1.0"
description = "AI-powered CI/CD Pipeline Orchestrator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "google-genai>=1.0.0",
    "gitpython>=3.1.40",
    "networkx>=3.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "sqlmodel>=0.0.14",
    "rich>=13.7.0",
    "anthropic>=0.40.0",
    "asyncpg>=0.29.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
